Emotion detection/requirements.txt
torch>=1.2.0
torchsummary
torchvision
numpy
Pillow
